(split)bug: none. move hardcoded cache limits and ttls out of provisioning and into localconfig.
Copied from Perforce
 Change: 21728

+++ src/java/com/zimbra/cs/account/ldap/LdapGroupEntryCache.java
@@ -60,13 +60,13 @@ class LdapGroupEntryCache {
     
     private LRUMap mIdCache;
     
-    private int mRefreshTTL;
+    private long mRefreshTTL;
 
 /**
  * @param maxItems
  * @param refreshTTL
  */
-    public LdapGroupEntryCache(int maxItems, int refreshTTL) {
+    public LdapGroupEntryCache(int maxItems, long refreshTTL) {
         mIdCache = new LRUMap(maxItems);
         mRefreshTTL = refreshTTL;
     }
